refactor(qpl): introduce trait to share question collection logic
